remove util class and duplicated tests
