+++ tck/src/main/java/ee/jakarta/tck/concurrent/common/fixed/counter/CounterRunnableTask.java
@@ -18,7 +18,6 @@ package ee.jakarta.tck.concurrent.common.fixed.counter;
 
 import java.time.Duration;
 
-import ee.jakarta.tck.concurrent.framework.TestUtil;
 import ee.jakarta.tck.concurrent.framework.junit.extensions.Wait;
 
 public class CounterRunnableTask implements Runnable, WorkInterface {
